tower: walker_common — one definition for the walks both tape walkers share (Phase B)
Post-split measurement found no whole section byte-identical between the
walkers (best pair 0.82): the "50% shared" mass lives in sub-blocks. The
eight blocks that ARE identical after name canonicalization — and are
exactly the format-defined walks where drift is the §G-2 bug class — now
live once in walker_common.rs (510 lines incl. docs):

  walk_wiring_gkr_core        the GKR transcript walk (~120 shared)
  merge_and_replay_chain      the merge-chain splice + replay assert
  census_levels_and_chain_rows  per-level census + duplex-model assert
  emit_ligerito_spine_walk    the spine walk (~120; 12 params — the run
                              is identical end-to-end with no narrower
                              waist, so the overage is the lesser evil)
  family_h_source_wires       the shv/value/rdp/gamma wire gather
  extend_const_coords / baked_inner_t_vals / prefix_product
                              the anchor-expect narrow waists

Deliberately NOT shared: the anchor-expect wholes (14-16 free variables,
no waist), the b3-census eprintln head (the REAL side interleaves its pow
census — sharing would reorder stderr), and every block whose two sides
differ for protocol reasons (lc-round tuple shapes, group-loop schedules,
el_rec vs piop_i). Each shared fn's doc names its walk and states both
walkers must stay on it.

Walkers: child 3,550 -> 3,208 lines, real 3,703 -> 3,359.

Verified: both tape pins after every extraction; ignored tower suite
13/13; full workspace suite 628 tests green; clippy -D warnings clean on
aarch64 and the CI x86_64 leg.
