+++ docs/bloat-ledger.md
@@ -464,7 +464,7 @@ alone: 23% clone-covered, ~2,970 of the recommended total.
 
 | # | Cluster | Dup lines | Consol. | Risk |
 | --- | --- | --- | --- | --- |
-| 1 | tower.rs Real/Child tape+region+node family (walkers share 76%, emitters 78%) | 2,488 | ~2,000 | **high** — transcript + byte pins; unify field-by-field, re-run pins **2026-09-03 re-scoped after measurement (Ron)**: LOC is not the point — measured 50% line-identical after name canonicalization (new 0.58 / emit 0.56 / check 0.22 similarity), and the walkers parse DIFFERENT inners (MixedInner leaf vs LeafOuter envelope), so a parameterized mega-walker would be flag soup. Revised plan: A0 pin the REAL tape (DONE, `real_tape_regions_pinned`), A section-split both monoliths along their `// ----` markers (pure motion, byte-pinned), B share ONLY the byte-identical format-defined sections (chain materials, spine quad replay, H(publics), recombination, jagged emission, ordinal walking), NO WalkSide parameterization, `check_*` never merged. The real fix is a **declarative op-schedule/region table** both walkers and the pins derive from — the same conclusion as cluster 9 — sequenced with tower productionization, not this campaign. |
+| 1 | tower.rs Real/Child tape+region+node family (walkers share 76%, emitters 78%) | 2,488 | ~2,000 | **high** — transcript + byte pins; unify field-by-field, re-run pins **2026-09-03 re-scoped after measurement (Ron)**: LOC is not the point — measured 50% line-identical after name canonicalization (new 0.58 / emit 0.56 / check 0.22 similarity), and the walkers parse DIFFERENT inners (MixedInner leaf vs LeafOuter envelope), so a parameterized mega-walker would be flag soup. Revised plan: A0 pin the REAL tape (DONE, `real_tape_regions_pinned`), A section-split both monoliths along their `// ----` markers (pure motion, byte-pinned), B share ONLY the byte-identical format-defined sections (chain materials, spine quad replay, H(publics), recombination, jagged emission, ordinal walking), NO WalkSide parameterization, `check_*` never merged. The real fix is a **declarative op-schedule/region table** both walkers and the pins derive from — the same conclusion as cluster 9 — sequenced with tower productionization, not this campaign. **A+B EXECUTED 2026-09-03**: A0 real-side pin + section split (13+8 / 8+10 named helpers, byte-pinned); B measurement post-split found NO whole section byte-identical (best pair 0.82) — the shareable truth is sub-blocks, extracted as `walker_common.rs` (8 wire-format walks, ~430 shared lines: the GKR transcript walk, chain merge/replay + level census, the spine walk, family-H sources, three anchor-expect waists); genuinely protocol-different pieces left per-walker on purpose. Cluster CLOSED; the op-schedule table remains the endgame. |
 | 2 | LegacyPow payload-ordinal walkers ×4 (2 of 4 had the bug) | 40 | 40 | high value, mechanical: one `payload_ordinals()` iterator — **DONE 2026-08-29**: `TranscriptOp::carries_payload()` replaces the 5 inline `ObserveBytes|Pow|LegacyPow` patterns (child/real walkers, fold_region, query) |
 | 3 | `GateType::witness()` 12-line body ×18 + `*256` twin gates | 420 | ~330 | low (witness) / medium (`*256` column order) — **DONE 2026-08-29** (witness half): `SlotWitness::element_from_rows` replaces the 14 identical bodies; the `*256` twin gates are left as they are (column order is protocol) |
 | 4 | tower native replicas (`sk_at_vks` ≡ `eval_sk_at_vks`, `frob_inv_native` ≡ jagged's) — export core fns instead | 700 | ~250 | low for replicas; do NOT merge the `check_*` validators (soundness backstop) — **DONE 2026-08-29** for the replicas: `sk_at_vks` → `ligerito::eval_sk_at_vks`, `frob_inv_native` → `jagged::frob_inv` (now pub; `frob_inv_matches_127_squarings` pins the two derivations); the `check_*` validators untouched as recommended |
